requirements.txt
yt-dlp
openai>=1.0.0
python-dotenv
tqdm
requests
pydub
ffmpeg
langchain-core
langchain
langchain-openai
langchain-text-splitters
langchain-pinecone
pinecone-client
langchain-google-genai
sentence-transformers
youtube-transcript-api
fastapi
uvicorn
pydantic
python-multipart